2.02 refactor ForecastFragment into to separate file
